Rename Navbar component file
